requirements.txt
pandas>=2.1.3,<2.3.0
numpy>=1.26.2,<2.0.0
scikit-learn>=1.3.2,<1.6.0
xgboost>=2.0.2,<3.0.0
lightgbm>=4.1.0,<5.0.0
transformers>=4.35.2,<4.50.0
torch>=2.2.0
streamlit>=1.28.2,<2.0.0
matplotlib>=3.8.2,<4.0.0
seaborn>=0.13.0,<0.14.0
plotly>=5.18.0,<6.0.0
nltk>=3.8.1,<4.0.0
textblob>=0.17.1,<0.19.0
wordcloud>=1.9.3,<2.0.0
google-api-python-client>=2.108.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
joblib>=1.3.2,<2.0.0
pytest>=7.4.0,<9.0.0
pydantic>=2.5.0,<3.0.0
pyyaml>=6.0.1,<7.0.0
fastapi>=0.104.0,<0.120.0
uvicorn>=0.24.0,<0.35.0
